Report TOML parsing errors
Without this, invalid TOML results in a JS error complaining that `in`
cannot be used to check if `template` is in `undefined`. This is because
bobadil signals invalid TOML by setting the `result` attribute to
`undefined`. With this change, the user gets somewhat more helpful
output, e.g.,

```
Errors when parsing TOML from file contract.toml: MismatchedTokenException: Expecting token of type --> CloseArray <-- but found --> '"' <--
```