Report the nature of latex errors
When tectonic fails, the current error reporting only gives the exit
code. This provides virtually no information about what went wrong. This
change prints the contents of stderr whenever tectonic fails, so that
the user will know what is wrong with the template. While it was
possible for the user to get this information by rerunning in verbose, I
think in this case, at this stage, the user will always want to know
what went wrong, so there's no point having them take the extra step of
running the command again with the verbose flag.

This output can sometimes be unsightly, but at this prototyping stage I
think the clearer signalling is well worth it.