Remove hardcoded font
I think it's best to just let pandoc use whatever font it defaults to.
This ensures that if pandoc works on the users machine for generating
PDFs, then it'll work to render our templates.

We can follow this up with changes to make the font configurable,
via #19, or some other means, to that users can specify Helvetica or
whatever else if they so choose.

This came up for me because contract compilation would fail due to not
having Helvetica installed.