Adds support for includes files to tweak latex
This change adds two default config files for adding latex styling to
the header of a template and prepending before the content. This won't
affect compiling directly from latex, since we use tectonic rather than
pandoc for this.

This also uses the new configurability to produce nicer formatting of
from markdown templates.

Signed-off-by: Shon Feder <shon@informal.systems>