Fix signatories bug
Without this fix, an error will be thrown every time we try to sign a
contract due to the `updateContractSignatories` method not being able to
interpret the data type of the "signatories" entry in the supplied
parameters. It seems a little extraneous to do all these type checks,
but it appears it's necessary.

Signed-off-by: Thane Thomson <connect@thanethomson.com>