Don't use "personal" as ID for signature example
It doesn't make sense to have a signature with the id "personal".
Signature id's are used to identify signatories, which means they must
not have duplicate IDs, but if everyone used the ID "personal" they
would all be clashing when trying to sign the same document. Also,
the distinction between profiles and signatures is meant to allow
different profiles to share the same signatures, so the example of the
"personal" profile using a signature with the ID "personal" doesn't make
sense.

Signed-off-by: Shon Feder <shon@informal.systems>