Add instructions on retroactive DCO signoff
Addressing a discussion that arose in #110.

After discussing internally I reasoned that the best way forward was to
provide more support for instructing contributors on how to
retroactively add DCO signoffs. This has the benefit of keeping to our
principles and of spreading knowledge, and encouraging contributors.
Thankfully, we also don't need to do the write up with instructions,
because `source{d}` already has a nice tutorial available. :)

Signed-off-by: Shon Feder <shon.feder@gmail.com>